+++ src/components/Shop/Common/Category/index.tsx
@@ -1,8 +1,9 @@
-import Link from 'next/link';
 import { useState, useEffect } from 'react';
+import Link from 'next/link';
 import Styled from './styles';
 import Plant from './Plant';
 import { getProductCategory } from 'src/apis/common/category';
+import { ICategory } from 'src/types/home/types';
 
 const Category = () => {
   const [categoryList, setCategoryList] = useState([]);
@@ -12,16 +13,17 @@ const Category = () => {
   };
   useEffect(() => {
     handleCategoryList();
-  }, [])
+  }, []);
   return (
     <Styled.Wrapper>
-      <Styled.Title >카테고리</Styled.Title>
+      <Styled.Title>카테고리</Styled.Title>
       <Styled.Plants>
-        {categoryList.map(plant => (
-          <Link href={`/shop/${plant.categoryName}`}  key={plant.categoryId}>
-            <Plant name={plant.categoryName} src={plant.imgUrl} />
-          </Link>
-        ))}
+        {categoryList &&
+          categoryList?.map((plant: ICategory) => (
+            <Link href={`/shop/${plant.categoryName}`} key={plant.categoryId}>
+              <Plant name={plant.categoryName} src={plant.imgUrl} />
+            </Link>
+          ))}
       </Styled.Plants>
     </Styled.Wrapper>
   );
